+++ Editor/Controls/BetterEnumFlagsField.cs
@@ -2,12 +2,14 @@ using System;
 using System.Reflection;
 using UnityEditor;
 using UnityEditor.UIElements;
+using UnityEngine;
+using UnityEngine.UIElements;
 
 namespace Vertx.Attributes.Editor
 {
-	public class BetterEnumFlagsField : EnumFlagsField
+	public class BetterEnumFlagsField : VisualElement
 	{
-		private EnumFlagsDrawer.ValueAndNames _valueAndNames;
+		private EnumFlagsValueAndNames _enumFlagsValueAndNames;
 
 		private delegate FieldInfo GetTypeFromPropertyBase(SerializedProperty property, out Type type);
 
@@ -24,14 +26,22 @@ namespace Vertx.Attributes.Editor
 			s_GetTypeFromProperty(property, out Type type);
 			return type;
 		}
-		
-		public BetterEnumFlagsField(SerializedProperty property, bool hideObsoleteNames) : base(property.displayName)
+
+		public BetterEnumFlagsField(SerializedProperty property, FieldInfo fieldInfo, bool hideObsoleteNames) : base()
 		{
-			bindingPath = property.propertyPath;
-			_valueAndNames = new EnumFlagsDrawer.ValueAndNames(GetTypeFromProperty(property), hideObsoleteNames);
-			// TODO actually add all these values.
-			/*choices = new List<string>(_valueAndNames.displayNames);
-			choicesMasks = new List<int>(m_EnumData.flagValues);*/
+			_enumFlagsValueAndNames = EnumFlagsValueAndNames.Get(fieldInfo, hideObsoleteNames);
+			var dropdownButton = new DropdownButton(property.displayName, _enumFlagsValueAndNames.GetName(property.intValue));
+			dropdownButton.RegisterClickCallback<(SerializedProperty property, EnumFlagsValueAndNames valuesAndNames)>(
+				(_, button, data) =>
+				{
+					Rect bounds = button.Q<VisualElement>(null, DropdownButton.DropdownUssClassName)?.worldBound
+					              ?? button.worldBound;
+					data.valuesAndNames.DropDown(bounds, data.property);
+				},
+				(property, _enumFlagsValueAndNames)
+			);
+			dropdownButton.TrackPropertyValue(property, p => dropdownButton.Text = _enumFlagsValueAndNames.GetName(p.intValue));
+			Add(dropdownButton);
 		}
 	}
 }