+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.5.0] - 2022-09-01
-- Added UIToolkit support to [HelpBox], [MinMax], [CurveDisplay], [File], [Directory].
+- Added UIToolkit support to all property drawers. Support may not be present in all Unity versions.
 - Added a Height property to CurveDisplay, allowing for larger display of the curve field.
+- Added [EnumDropdown], this shows an AdvancedDropdown instead of a GenericMenu, allowing for scrolling in long enums.
 
 ## [1.4.1] - 2022-08-28
 - Added Relabel drawer.